requirements.txt
Django==2.2.17
django-adminlte2==0.4.1
django-stdimage==5.0.1
et-xmlfile==1.0.1
jdcal==1.4.1
mysql-connector-python==8.0.23
openpyxl==3.0.6
pillow==8.1.0
progressbar2==3.43.1
protobuf==3.15.0
psycopg2==2.8.6
python-utils==2.3.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
